# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking tables in restaurants API",
        "description": "API documentation for the restaurant booking system.",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Получение списка пользователей",
                "description": "Получение списка всех пользователей. Доступно только администраторам.",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Поиск по имени",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Поиск по email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Поиск по phone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_blocked",
                        "in": "query",
                        "description": "Фильтрация по статусу блокировки",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "false",
                            "enum": [
                                "true",
                                "false"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение списка пользователей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Массив пользователей",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Илларион Иванович Шарапов"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "admin@admin.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+590432015354"
                                                    },
                                                    "is_blocked": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "13.10.2025 16:58:09"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "13.10.2025 16:58:09"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "description": "Пагинация",
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 56
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "count_pages": {
                                                    "type": "integer",
                                                    "example": 6
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Вы не авторизованы",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "https://example.com/errors/unauthorized"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "You not authorized"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "Доступ к ресурсу доступен только авторизованным пользователям!"
                                        },
                                        "instance": {
                                            "type": "string",
                                            "example": "/api/users"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Вы не авторизованы",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "https://example.com/errors/forbidden"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "You not authorized"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 403
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "Доступ к ресурсу запрещен!"
                                        },
                                        "instance": {
                                            "type": "string",
                                            "example": "/api/users"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}